requirements.txt
colorlog==6.9.0
dnspython==2.7.0
pymongo==4.12.0
python-dotenv==1.1.0
tqdm==4.67.1
